api/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
python-multipart
opencv-python-headless

